make: use deferred conditional var operator to assign defaults
This allows the host environment to easily override variables like
PREFIX, LDFLAGS and CC by setting the environment variables. A lot of
distro packaging tools set those in the build environment and expect
the build target to simply pick those instead of forcefully overwriting
them.

Use the same easy operator for providing base defaults for VERSION and
PREFIX as well.